#43: Remove unused handles, code, fixmes
