+++ src/models/chart.model.ts
@@ -1,0 +1,12 @@
+
+export interface VerticalChartLine {
+  type: string;
+  label: string;
+  value: number;
+  color: string;
+}
+
+export interface SleepChartBox {
+  valueStart: number;
+  valueEnd: number;
+}
