+++ src/modules/bulk-updates/bulk-updates.component.ts
@@ -60,7 +60,7 @@ export class BulkUpdatesComponent implements OnInit {
 
     this.updateService.mo.c8y_Series = this.simSettings.allInstructionsArray;
     this.updateService.mo.c8y_mirroredAxis = this.mirroredAxis;
-    this.updateService.updateSimulatorObject(this.updateService.mo).then((res) => {});
+    this.updateService.updateSimulatorObject(this.updateService.mo);
   }
 
   toggleIntertwineSeries() {
@@ -163,7 +163,7 @@ export class BulkUpdatesComponent implements OnInit {
 
     this.updateService.mo.c8y_Series = this.simSettings.allInstructionsArray;
     this.updateService.mo.c8y_saltValue = this.saltValue;
-    this.updateService.updateSimulatorObject(this.updateService.mo).then((res) => {});
+    this.updateService.updateSimulatorObject(this.updateService.mo);
   }
 
   calculateTheDeviation(entry: IndexedCommandQueueEntry) {
