+++ src/modules/create-sim/sim-settings/sim-settings.component.ts
@@ -189,8 +189,6 @@ export class SimSettingsComponent {
     this.isSmartRestSelected = value === InstructionCategory.SmartRest;
   }
 
-  updateSeriesColor() {} // FIXME obsolete?
-
   openSimulatorUploadFileDialog() {
     const modal = this.modalService.show(SimulatorFileUploadDialog);
     this.subscriptions.push(
