+++ src/modules/create-sim/sim-settings/simulator-file-upload-dialog.ts
@@ -159,7 +159,7 @@ export class SimulatorFileUploadDialog {
 
       reader.onload = (e) => {
         const text = reader.result.toString();
-        resolve(text);
+        resolve(text); // FIXME: there is no reject for error case
       };
 
       reader.readAsText(file);
