+++ src/modules/series-item/series-list/series-list.component.ts
@@ -64,6 +64,6 @@ export class SeriesListComponent {
     this.instructionsSeries = updatedInstructionsSeries;
     this.simSettingsService.updateCommandQueueAndIndicesFromIndexedCommandQueue(this.indexedCommandQueue);
     this.simSettingsService.setAllInstructionsSeries(updatedInstructionsSeries);
-    this.updatedService.updateSimulatorObject(this.updatedService.mo).then((res) => {}); // FIXME proper handling
+    this.updatedService.updateSimulatorObject(this.updatedService.mo);
   }
 }
